Move status enum to hook file and add comment
